% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_compression_stats}
\alias{calc_compression_stats}
\title{Calculate ompression states}
\usage{
calc_compression_stats(init_size, final_size, prefix, verbosity)
}
\arguments{
\item{init_size, final_size}{file sizes}

\item{prefix}{message prefix}

\item{verbosity}{verbosity level. integer}
}
\description{
Calculate ompression states
}
