% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_exe}
\alias{check_exe}
\title{Check the Executable exists}
\usage{
check_exe(prog_name, verbosity = 0)
}
\arguments{
\item{prog_name}{name of program to search for}

\item{verbosity}{verbosity level. default 0}
}
\value{
path
}
\description{
Check the Executable exists
}
