% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_filename}
\alias{check_filename}
\title{Check that a filename is really a filename}
\usage{
check_filename(filename)
}
\arguments{
\item{filename}{single filename to check}
}
\value{
TRUE if all checks pass, otherwise throws an ERROR
}
\description{
Lots of sanity checks to avoid mischief by the user: Must be single
character string with nchars > 0.  File must exist.
}
\details{
This is still open to race conditions.
}
