% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpg.R
\name{jpegoptim}
\alias{jpegoptim}
\title{Lossless and Lossy compression of JPEG files}
\usage{
jpegoptim(infile, quality = NULL, size = NULL, outfile = infile, verbosity = 0)
}
\arguments{
\item{infile}{PNG file to optimise}

\item{quality}{set maximum image quality factor. Valid quality values: 0 - 100.
Default: NULL (not enabled)}

\item{size}{Try to optimize file to given size. Target size is specified
either in kilo bytes (1 - n) or as percentage (1\% - 99\%)}

\item{outfile}{output PNG location. If not specified, input file will be
overwritten}

\item{verbosity}{verbosity level. Possible values: 0, 1, 2. Default: 0}
}
\description{
If no \code{outfile} is specified, then the input file is compressed in place.
}
\details{
If both \code{quality} and \code{size} options are left at their default of
\code{NULL}, then lossless optimization is performed.
}
