% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png-optipng.R
\name{optipng}
\alias{optipng}
\title{Lossless compression/optimization of PNG files}
\usage{
optipng(infile, level = 1, outfile = infile, verbosity = 0)
}
\arguments{
\item{infile}{PNG file to optimise}

\item{level}{Compression level. Valid values 0-7. Default 1.}

\item{outfile}{output PNG location. If not specified, input file will be
overwritten}

\item{verbosity}{verbosity level. Possible values: 0, 1, 2. Default: 0}
}
\description{
If no \code{outfile} is specified, then the input file is compressed in place.
}
