% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{pdfopt}
\alias{pdfopt}
\title{Lossy compression of PDF files}
\usage{
pdfopt(
  infile,
  quality = c("screen", "ebook", "prepress", "printer", "default"),
  dpi = 72,
  outfile = infile,
  verbosity = 0
)
}
\arguments{
\item{infile}{PNG file to optimise}

\item{quality}{character. One of \code{c('screen', 'ebook', 'prepress', 'printer', 'default')}
default: 'screen'}

\item{dpi}{Embedded image resolution. Default: 72}

\item{outfile}{output PNG location. If not specified, input file will be
overwritten}

\item{verbosity}{verbosity level. Possible values: 0, 1, 2. Default: 0}
}
\description{
If no \code{outfile} is specified, then the input file is compressed in place.
}
\details{
\code{gs} options taken from \url{http://www.alfredklomp.com/programming/shrinkpdf/}
}
