% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png-pngcrush.R
\name{pngcrush}
\alias{pngcrush}
\title{Lossless compression/optimization of PNG files}
\usage{
pngcrush(infile, brute = FALSE, outfile = infile, verbosity = 0)
}
\arguments{
\item{infile}{PNG file to optimise}

\item{brute}{use brute-force: try 176 different methods. Default: FALSE}

\item{outfile}{output PNG location. If not specified, input file will be
overwritten}

\item{verbosity}{verbosity level. Possible values: 0, 1, 2. Default: 0}
}
\description{
If no \code{outfile} is specified, then the input file is compressed in place.
}
