% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png-pngquant.R
\name{pngquant}
\alias{pngquant}
\title{Lossless compression/optimization of PNG files}
\usage{
pngquant(infile, speed = 4, dither = FALSE, outfile = infile, verbosity = 0)
}
\arguments{
\item{infile}{PNG file to optimise}

\item{speed}{speed/quality trade-off. 1=slow, 4=default, 11=fast & rough. Default: 4}

\item{dither}{logical. Use dithering? Default: FALSE}

\item{outfile}{output PNG location. If not specified, input file will be
overwritten}

\item{verbosity}{verbosity level. Possible values: 0, 1, 2. Default: 0}
}
\description{
If no \code{outfile} is specified, then the input file is compressed in place.
}
