% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shwhich}
\alias{shwhich}
\title{The equivalent of running 'which prog_name' at the shell}
\usage{
shwhich(prog_name)
}
\arguments{
\item{prog_name}{name of program to search for}
}
\value{
path to file, or NULL if not found
}
\description{
The equivalent of running 'which prog_name' at the shell
}
