% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png-zopflipng.R
\name{zopflipng}
\alias{zopflipng}
\title{Lossless compression/optimization of PNG files}
\usage{
zopflipng(
  infile,
  lossy_alpha = FALSE,
  lossy_8bit = FALSE,
  more = FALSE,
  insane = FALSE,
  outfile = infile,
  verbosity = 0
)
}
\arguments{
\item{infile}{PNG file to optimise}

\item{lossy_alpha}{remove colors behind alpha channel 0. No visual difference, removes hidden information.}

\item{lossy_8bit}{convert 16-bit per channel image to 8-bit per channel.}

\item{more}{compress more? default: FALSE}

\item{insane}{insane compression options. default: FALSE}

\item{outfile}{output PNG location. If not specified, input file will be
overwritten}

\item{verbosity}{verbosity level. Possible values: 0, 1, 2. Default: 0}
}
\description{
If no \code{outfile} is specified, then the input file is compressed in place.
}
